Remove edges to anything other than root for inputs
If we can provide a value, then that value should not be sourced from
anywhere else. By keeping the other edges (that we'd never use), we get
cycles when we have bidirectional converters.

For example, if you can convert from string to []byte and []byte to
string, this creates a cycle. But if we have a string available, then we
can ignore the fact we could in theory get that string from []byte. We
do this before we can calculate shortest path since our shortest path
algorithm requires acyclic graphs up front (currently using Kahn's
algoritm).